+++ src/lib/libast/misc/recfmt.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -60,9 +60,9 @@ recfmt(const void* buf, size_t size, off_t total)
 
 	s = (unsigned char*)buf;
 	t = s + size;
-	while ((k = (t - s)) >= 4 && !s[2] && !s[3])
+	while ((k = (unsigned int)(t - s)) >= 4 && !s[2] && !s[3])
 	{
-		if ((i = (s[0]<<8)|s[1]) > k)
+		if ((i = (unsigned)((s[0]<<8)|s[1])) > k)
 			break;
 		s += i;
 	}
@@ -75,7 +75,7 @@ recfmt(const void* buf, size_t size, off_t total)
 	 */
 
 	for (i = 0; i < elementsof(terminators); i++)
-		if ((t = (unsigned char*)memchr(s, k = terminators[i], size / 2)) && (n = t - s + 1) > 1 && (total <= 0 || !(total % n)))
+		if ((t = (unsigned char*)memchr(s, k = terminators[i], size / 2)) && (n = (unsigned int)(t - s + 1)) > 1 && (total <= 0 || !(total % n)))
 		{
 			for (j = n - 1; j < size; j += n)
 				if (s[j] != k)
@@ -108,7 +108,7 @@ recfmt(const void* buf, size_t size, off_t total)
 	}
 	n = 0;
 	m = 0;
-	f = ~0;
+	f = ~0UL;
 	for (i = x; i > 1; i--)
 	{
 		if ((total <= 0 || !(total % i)) && q->rep[i] > q->rep[n])
@@ -134,7 +134,7 @@ recfmt(const void* buf, size_t size, off_t total)
 			for (j = 0; j < elementsof(terminators); j++)
 				if (s[i] == terminators[j])
 					n++;
-		n = n ? 0 : total;
+		n = n ? 0 : (unsigned int)total;
 	}
 	free(q);
 	return n ? REC_F_TYPE(n) : REC_N_TYPE();
